// src/internal.d.ts
export type ComponentChild =
  | VNode<any>
  | string
  | number
  | boolean
  | null
  | undefined;
export type ComponentChildren = ComponentChild[] | ComponentChild;

export interface FunctionComponent<P = {}> extends preact.FunctionComponent<P> {
  // Internally, createContext uses `contextType` on a Function component to
  // implement the Consumer component
  contextType?: PreactContext;

  // Internally, createContext stores a ref to the context object on the Provider
  // Function component to help devtools
  _contextRef?: PreactContext;

  // Define these properties as undefined on FunctionComponent to get rid of
  // some errors in `diff()`
  getDerivedStateFromProps?: undefined;
  getDerivedStateFromError?: undefined;
}

export interface ComponentClass<P = {}> extends preact.ComponentClass<P> {
  _contextRef?: any;

  // Override public contextType with internal PreactContext type
  contextType?: PreactContext;
}

// Redefine ComponentType using our new internal FunctionComponent interface above
export type ComponentType<P = {}> = ComponentClass<P> | FunctionComponent<P>;

export interface PreactElement extends HTMLElement {
  _children?: VNode<any> | null;
  /** Event listeners to support event delegation */
  _listeners?: Record<string, (e: Event) => void>;

  // Preact uses this attribute to detect SVG nodes
  ownerSVGElement?: SVGElement | null;

  // style: HTMLElement["style"]; // From HTMLElement

  data?: string | number; // From Text node
}

// We use the `current` property to differentiate between the two kinds of Refs so
// internally we'll define `current` on both to make TypeScript happy
type RefObject<T> = { current: T | null };
type RefCallback<T> = { (instance: T | null): void; current: undefined };
type Ref<T> = RefObject<T> | RefCallback<T>;

export interface VNode<P = Record<never, never>> extends preact.VNode<P> {
  // Redefine type here using our internal ComponentType type
  type: string | ComponentType<P>;
  props: P & { children: ComponentChildren };
  ref?: Ref<unknown> | null;
  _children: VNode<unknown>[] | null;
  _parent: VNode | null;
  _depth: number | null;
  /**
   * The [first (for Fragments)] DOM child of a VNode
   */
  _dom: PreactElement | null;
  /**
   * The last dom child of a Fragment, or components that return a Fragment
   */
  _nextDom: PreactElement | null;
  _component: Component | null;
  _hydrating: boolean | null;
  constructor: undefined;
  _original: number;
}

export interface Component<P = {}, S = {}> extends preact.Component<P, S> {
  // When component is functional component, this is reset to functional component
  constructor: ComponentType<P>;
  state: S; // Override Component["state"] to not be readonly for internal use, specifically Hooks
  base?: PreactElement;

  _dirty: boolean;
  _force?: boolean;
  _renderCallbacks: Array<() => void>; // Only class components
  _globalContext?: any;
  _vnode?: VNode<P> | null;
  _nextState?: S | null; // Only class components
  /** Only used in the devtools to later dirty check if state has changed */
  _prevState?: S | null;
  /**
   * Pointer to the parent dom node. This is only needed for top-level Fragment
   * components or array returns.
   */
  _parentDom?: PreactElement | null;
  // Always read, set only when handling error
  _processingException?: Component<any, any> | null;
  // Always read, set only when handling error. This is used to indicate at diffTime to set _processingException
  _pendingError?: Component<any, any> | null;
}